% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_search.R
\name{AIC_smooth_spline}
\alias{AIC_smooth_spline}
\title{Calculate the Akaike Information Criterion for a smooth.spline}
\usage{
AIC_smooth_spline(fittedSmoothSpline)
}
\arguments{
\item{fittedSmoothSpline}{A fitted \code{\link[stats]{smooth.spline}}}
}
\value{
The AIC value.
}
\description{
Calculate the Akaike Information Criterion (\emph{AIC}) for a fitted \code{\link[stats]{smooth.spline}}. The smaller the AIC, the better the spline fit.
}
