% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_search.R
\name{AICc_smooth_spline}
\alias{AICc_smooth_spline}
\title{Calculate the Akaike Information Criterion Corrected for small observation numbers for a smooth.spline}
\usage{
AICc_smooth_spline(fittedSmoothSpline)
}
\arguments{
\item{fittedSmoothSpline}{A fitted \code{\link[stats]{smooth.spline}}}
}
\value{
The AICc value.
}
\description{
Calculate the Akaike Information Criterion Corrected for small observation numbers (\emph{AICc}) for a fitted \code{\link[stats]{smooth.spline}}. The smaller the AICc, the better the spline fit.
}
