% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_search.R
\name{loglik_smooth_spline}
\alias{loglik_smooth_spline}
\title{Calculate the penalised loglikelihood of a smooth.spline}
\usage{
loglik_smooth_spline(fittedSmoothSpline)
}
\arguments{
\item{fittedSmoothSpline}{A fitted \code{\link[stats]{smooth.spline}}}
}
\value{
The penalised loglikelihood.
}
\description{
Calculate the penalised loglikelihood of a \code{\link[stats]{smooth.spline}} using the integrated second derivative. The likelihood consists of 1) the (weighted) residuals sum of squares, 2) a penalty term (integrated second derivative = total curvature). The smaller the penalised loglikelihood, the better the fit as the residuals and penalty on roughness are minimised. Adapted from \code{aroma.light::likelihood.smooth.spline}.
}
