% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/santaR.R
\docType{package}
\name{santaR}
\alias{santaR-package}
\alias{santaR}
\alias{SANTAR}
\title{santaR: A package for Short AsyNchronous Time-series Analysis in R}
\description{
\pkg{santaR} provides a graphical and automated pipeline for the analysis of short time-series studies.
It enables the detection of significantly altered time trajectories between study groups, while
being resilient to missing values and unsynchronised measurements.
}
\details{
The main functions of \pkg{santaR} are \code{\link{santaR_start_GUI}} to start the graphical user
interface, as well as \code{\link{santaR_auto_fit}} and \code{\link{santaR_auto_summary}} for
automated command line analysis and reporting. Refer to the vignettes for graphical user
interface and command line tutorials.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/adwolfer/santaR}
  \item \url{https://adwolfer.github.io/santaR/}
  \item Report bugs at \url{https://github.com/adwolfer/santaR/issues/new}
}

}
\author{
\strong{Maintainer}: Arnaud Wolfer \email{adwolfer@gmail.com} (\href{https://orcid.org/0000-0001-5856-3218}{ORCID})

Other contributors:
\itemize{
  \item Timothy Ebbels \email{t.ebbels@imperial.ac.uk} [contributor]
  \item Joe Cheng \email{joe@rstudio.com} (Shiny javascript custom-input control) [contributor]
}

}
